+++ build_runner/test/integration_tests/test_command_test.dart
@@ -22,21 +22,37 @@ void main() async {
     );
 
     // `test` does not support specifying directory to build.
-    await tester.run(
+    var output = await tester.run(
       'root_pkg',
       'dart run build_runner test --force-jit web',
       expectExitCode: ExitCode.usage.code,
     );
-    await tester.run(
+    expect(
+      output,
+      contains(
+        'The `test` command requires a `--` separator '
+        'before any test arguments.',
+      ),
+    );
+
+    output = await tester.run(
       'root_pkg',
       'dart run build_runner test --force-jit web -- -p chrome',
       expectExitCode: ExitCode.usage.code,
     );
+    expect(
+      output,
+      contains(
+        'The `test` command requires a `--` separator '
+        'before any test arguments.',
+      ),
+    );
 
-    // Requires `build_test` dependency.
-    final output = await tester.run(
+    // Corrected command instead fails with config error because `build_test`
+    // dependency is missing.
+    output = await tester.run(
       'root_pkg',
-      'dart run build_runner test --force-jit',
+      'dart run build_runner test --force-jit -- web',
       expectExitCode: ExitCode.config.code,
     );
     expect(
@@ -46,5 +62,26 @@ void main() async {
         'which is required to run tests.',
       ),
     );
+
+    // Add the `build_test` dependency, the argument is passed to `test` as
+    // expected.
+    tester.writePackage(
+      name: 'root_pkg',
+      dependencies: ['build_runner', 'build_test'],
+      files: {
+        'test/passes_fails_test.dart': '''
+import 'package:test/test.dart';
+void main() {
+  test('passes', () {});
+  test('fails', () {
+    fail('should not run');
+  });
+}
+''',
+      },
+    );
+
+    // Succeds because only 'passes' runs.
+    await tester.run('root_pkg', 'dart run build_runner test -- -n passes');
   });
 }